Removed UpdateImportantChannels from OnShutdown because there is already an update done by OnStatusNameChange
Added IrcCmdPrefix Char to config.
Added CommandHandler.RemoteCommandPrefix to OnText
Then it is set each time there is text for command changes.

Removed extra check from UpdateTopic so now it only looks for
AutomaticTopicUpdating == true and continues to execute.

Put conditions for if statment on 1 line, to make it a bit easier.

Removed old else, and added 2 new if functions 1 for shutdown and 1 for Locked

Fixes Issue:5